requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
boto3==1.24.46
botocore==1.27.46
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
idna==3.3
jmespath==1.0.1
kombu==5.2.4
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
PyJWT==2.4.0
pymemcache==3.5.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
yarg==0.1.9
